server/requirements.txt
astroid==2.4.2
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cycler==0.10.0
dataclasses==0.6
Flask==1.1.2
Flask-Cors==3.0.9
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
lxml==4.5.1
MarkupSafe==1.1.1
mccabe==0.6.1
pathtools==0.1.2
praw==7.1.0
prawcore==1.5.0
psycopg2-binary==2.8.6
pylint==2.5.3
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.14.0
requests==2.24.0
SQLAlchemy==1.3.19
toml==0.10.1
typing==3.7.4.1
update-checker==0.18.0
urllib3==1.25.9
Werkzeug==1.0.1
wrapt==1.12.1
